docs(scalars): add stories for the toggle
