+++ packages/design-system/src/scalars/components/boolean-field/boolean-field.stories.tsx
@@ -57,7 +57,7 @@ const meta = {
     docs: {
       description: {
         component:
-          "A boolean field component that can be used as a [checkbox](?path=/docs/document-engineering-data-entry-checkbox--readme) or [toggle](?path=/docs/document-engineering-fragments-toggle-field--readme) depending on the `isToggle` prop.",
+          "A boolean field component that can be used as a [checkbox](?path=/docs/document-engineering-data-entry-checkbox--readme) or [toggle](?path=/docs/document-engineering-data-entry-toggle--readme) depending on the `isToggle` prop.",
       },
     },
   },
