+++ packages/design-system/src/scalars/components/fragments/index.ts
@@ -7,4 +7,3 @@ export * from "./radio-group-field/index.js";
 export * from "./select-field/index.js";
 export * from "./text-field/index.js";
 export * from "./textarea-field/index.js";
-export * from "./toggle-field/index.js";
